refactor: 💡 Sources
Move Python files into one sources folder
